% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_count_compilation.R
\name{cell_count_compilation}
\alias{cell_count_compilation}
\title{Compile cell counts from multiple brains.}
\usage{
cell_count_compilation(
  regions = c("grey"),
  subregions = 5,
  clean_zeros = FALSE,
  hierarchy_level = FALSE,
  datasets = c(),
  files_list = c(),
  brains_list = c(),
  output = NULL
)
}
\arguments{
\item{regions}{(optional, default = c("grey")) Vector of parent regions.}

\item{subregions}{(optional, default = 5) Number of hierarchical levels
below parent regions to include cell counts for.}

\item{clean_zeros}{(optional, default = FALSE) Boolean to specify whether
regions with 0 cells in all brains are removed.}

\item{datasets}{(required, default = c()) Vector of dataset names (in quotes)
corresponding to the relevant dataset in each .RData file.}

\item{files_list}{(required, default = c()) Vector of .RData files (including
file paths) to be accessed.}

\item{brains_list}{(required, default = c()) Vector of brain names (to be used as
column headers).}

\item{output}{(required, default = NULL) Path to .csv file to be created.}
}
\value{
Returns an object containing regional cell counts from
all specified datasets.
}
\description{
Loop through datasets across multiple .RData files,
corresponding to different SMART analyses, and compile cell counts
from each of them into a data table.
}
