% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_counter.R
\name{cell_counter}
\alias{cell_counter}
\title{Plate-by-plate cell counter for \emph{segs} object.}
\usage{
cell_counter(segs = segs)
}
\arguments{
\item{segs}{(required) Segmentation data.}
}
\description{
Loop through all segmented cells to determine how
many cells were segmented, removed by \code{\link[=clean_duplicates]{clean_duplicates()}}, and
retained after duplicate cleanup. Prints output in console.
}
