% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenate.R
\name{concatenate}
\alias{concatenate}
\title{Concatenate multiple datasets}
\usage{
concatenate(data_list = c(), datasets = c())
}
\arguments{
\item{data_list}{(required, default = c()) Vector of .RData files containing
datasets to be combined.}

\item{datasets}{(required, default = c()) Vector of dataset names, in quotes,
corresponding to each .RData file}
}
\value{
Returns \emph{group_dataset} a variable containing cells from all initial
datasets.
}
\description{
Loop through datasets across multiple .RData files,
corresponding to different SMART analyses, and combine datasets to form
one master dataset
}
