% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_groups.R
\name{get_groups}
\alias{get_groups}
\title{Compile group data from individual brains.}
\usage{
get_groups(
  input = NULL,
  brains_list = c(),
  groups_list = c(),
  groups = c(),
  output = NULL
)
}
\arguments{
\item{input}{(required, default = NULL) Path to .csv file to be used,
the output of \code{\link[=cell_count_compilation]{cell_count_compilation()}}.}

\item{brains_list}{(required, default = c()) Vector of brain names (column
headers in .csv file).}

\item{groups_list}{(required, default = c()) Vector corresponding to brains_list
that specifies the group name for each brain.}

\item{groups}{(required, default = c()) Vector of groups.}

\item{output}{(required, default = NULL) Path to .csv file to be created.}
}
\value{
Returns an object containing regional cell count means and
standard deviations for each test group.
}
\description{
Uses output of \code{\link[=cell_count_compilation]{cell_count_compilation()}} to calculate
the mean and standard deviation of every user-specified test group.
}
