% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isolate_dataset.R
\name{isolate_dataset}
\alias{isolate_dataset}
\title{Selectively isolate forward-warped cells \strong{(W)}.}
\usage{
isolate_dataset(
  setup,
  dataset,
  start_AP = NULL,
  end_AP = NULL,
  hemisphere = "B",
  plates = c(1:length(unique(dataset$image))),
  mirror = FALSE,
  flip = FALSE,
  bounds = c(),
  rois = c("grey")
)
}
\arguments{
\item{setup}{(required) Setup list from \code{\link[=setup_pl]{setup_pl()}}.}

\item{dataset}{(required) Dataset output from \code{\link[=forward_warp]{forward_warp()}}.}

\item{start_AP}{(optional, default = NULL) Most anterior AP to retain
cells.}

\item{end_AP}{(optional, default = NULL) Most posterior AP to retain
cells.}

\item{hemisphere}{(optional, default = "B") Hemisphere to retain cells
from ("L", "R", or "B").}

\item{plates}{(optional, default = c(1:length(unique(dataset$image))))
Vector of registration plates to retain plates from; specify plate numbers
relative to the first registration plate in the current analysis.}

\item{mirror}{(optional, default = FALSE) Boolean to specify whether all
cells will be mirrored across the midline.}

\item{flip}{(optional, default = FALSE) Boolean to specify whether all cells
will be flipped across the midline.}

\item{bounds}{(optional, default = c()) Vector to specify whether cells within
a circular or rectangular area should be retained. For circles, the vector
should be formatted as: c("circle", ML-center, DV-center, radius). For
rectangles, the vector should be formatted as: c("rectangle",
ML-top-left-corner, DV-top-left-corner, width, height).}

\item{rois}{(optional, default = c("grey")) Vector to specify regions of interest
to retain cells from.}
}
\value{
Returns \emph{isolated_dataset} a variable storing all cells retained from
\emph{dataset}.
}
\description{
put description here
}
