% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voxel_stats.R
\name{voxel_stats}
\alias{voxel_stats}
\title{Run statistical tests on voxel-based heatmaps}
\usage{
voxel_stats(
  input,
  group_matrices,
  groups = c(),
  ML_bounds = c(-5, 5),
  DV_bounds = c(-8, 1),
  p_value = 0.05,
  output
)
}
\arguments{
\item{input}{(required) Specify path to input .RData file containing the
group matrix.}

\item{group_matrices}{(required) Specify object name for matrix of groups.}

\item{groups}{(required, default = c()) Vector of 2 groups to be compared.}

\item{ML_bounds}{(optional, default = c(-5, 5)) Bounds for ML axis.}

\item{DV_bounds}{(optional, default = c(-8, 1)) Bounds for DV axis.}

\item{p_value}{(optional, default = 0.05) P-value threshold for significance.}

\item{output}{(required) Specify path to output folder to save heatmap and .RData output.}
}
\value{
Returns \emph{stat_matrix} an object containing a matrix with statistical test
output (-1 for significantly lower, 0 for no significance, 1 for significantly higher).
}
\description{
Generate statistical outputs from voxel-based heatmaps created
by \code{\link[=voxelize]{voxelize()}}. Statistical output resembles input heatmaps, with green/red
colors specifying locations where voxels show statistically significant
differences between groups.
}
