% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voxelize.R
\name{voxelize}
\alias{voxelize}
\title{Generate voxel-based heatmaps}
\usage{
voxelize(
  brains_list = c(),
  data_list = c(),
  datasets = c(),
  groups_list = c(),
  groups = c(),
  ML_bounds = c(-5, 5),
  DV_bounds = c(-8, 1),
  detection_bounds = c(0.1, 0.1),
  resolution = 50,
  heatmaps = TRUE,
  save = TRUE,
  output = NULL
)
}
\arguments{
\item{data_list}{(required, default = c()) Vector of .RData files (including
file paths) to be accessed.}

\item{datasets}{(required, default = c()) Vector of dataset names (in quotes)
corresponding to the relevant dataset in each .RData file. To generate heatmaps
for individual brain slices, use \code{\link[=isolate_dataset]{isolate_dataset()}} to isolate cells from a
specific plate.}

\item{groups_list}{(required, default = c()) Vector corresponding to brains_list
that specifies the group name for each brain.}

\item{groups}{(required, default = c()) Vector of groups.}

\item{ML_bounds}{(optional, default = c(-5, 5)) Bounds for ML axis.}

\item{DV_bounds}{(optional, default = c(-8, 1)) Bounds for DV axis.}

\item{detection_bounds}{(optional, default = c(0.1, 0.1)) Side length (in mm) of
detection square around each search point.}

\item{resolution}{(optional, default = 50) Number of search points in each dimension,
per millimeter (e.g. resolution 50 divides each square mm into a 50 x 50 grid with
2,500 search points).}

\item{heatmaps}{(optional, default = TRUE) Specify whether to output heatmaps.}

\item{save}{(optional, default = TRUE) Specify whether to save .csv files.}

\item{output}{(required, default = NULL) Specify path to output folder to output
.csv files, heatmaps, and .RData files.}
}
\value{
Returns \emph{group_matrices} an object containing all brain matrices, as well
as group mean and SD matrices.
}
\description{
Generate voxel-based cell density heatmaps based on datasets
from SMART analysis. Generate group mean and SD heatmaps to show trends across
test groups. Generate .csv files containing raw heatmap data for downstream
statistical analysis.
}
